refactor(orchestration): decompose 1262-line module into domain package
Split orchestration.py into orchestration/ package with 5 submodules:
- _helpers.py: scheduling and retraining utilities
- feature.py: feature pipeline orchestration
- training.py: training pipeline orchestration
- inference.py: inference pipeline step
- drift.py: drift detection steps

__init__.py re-exports all public/private symbols for full backward
compatibility. Tests updated to monkeypatch correct submodule bindings.
